requirements.txt
shortuuid==1.0.1
Flask==2.0.1
werkzeug==2.0.3
wtforms==2.3.3
flask-wtf==1.0.0
python-dotenv==0.19.0
alembic==1.6.5
SQLAlchemy==1.4.23
gunicorn==20.1.0
waitress==2.1.2